Added some utility functions to convert between UTF encodings. Changed the Win32 port to assume UTF8 strings for console output   and window titles and such. Fixed a bug where WGL extensions weren't always initialized properly. Removed the unused DirectInput keyboard handling code. Augmented configure to check the sizeof(long). This is used   to determine whether the platform is a 32 or 64 bit environment. Added types to cstypes.h which are (un)signed integers with the   same size as a pointer. They are intended in existing cases where   pointers are casted to ints/longs (it appears that long on 64-bit   MSVC is still 32 bits large, so it can't be used. Tho I'm not sure.) Augmented configure to check the sizeof(wchar_t).
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@14826 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/cssys/win32/registrycfg.h
@@ -62,15 +62,15 @@ private:
   struct Block_O_Mem
   {
     BYTE* data;
-    DWORD size;
+    size_t size;
     Block_O_Mem () : data(0), size(0) { }
     void Clear() { delete[] data; }
-    void SetSize (DWORD sz) { Clear(); size = sz; data = new BYTE[sz]; }
+    void SetSize (size_t sz) { Clear(); size = sz; data = new BYTE[sz]; }
     ~Block_O_Mem() { Clear(); }
   };
   // To shorten calls to RegSetValueEx()
   bool InternalSetValue (const char* Key,
-    DWORD type, const void* data, DWORD datasize);
+    DWORD type, const void* data, size_t datasize);
   // To shorten calls to RegQueryValueEx()
   bool InternalGetValue (const char* Key,
     DWORD& type, Block_O_Mem& data) const;
